moved all css styles into a "styles" folder
