+++ client/src/pages/PageNotFound.jsx
@@ -1,7 +1,7 @@
+import { Link as UILink } from "@nextui-org/react";
 import React from "react";
 import { Link } from "react-router-dom";
-import { Link as UILink } from "@nextui-org/react";
-import "./PageNotFound.css";
+import "../styles/PageNotFound.css";
 
 const PageNotFound = () => {
   return (
