+++ client/src/styles/PageNotFound.css
@@ -19,6 +19,7 @@
   display: inline-block;
   padding-right: 12px;
   animation: type 0.6s alternate infinite;
+  font-weight: bold;
 }
 
 @keyframes type {
